refactor: useFullscreenMode hook
